Cleanup, move modules into specific package.
The reason for this is because we can now do `CompileTime.include(ufront.ufadmin.modules)`
and it will automatically import all modules, even if they are never referenced.

This makes it easy for us to get a list of default modules to display.

I've also switched to using CompileTime.readFile() to get a template, and ViewResult to
render the template, rather than CompileTime.interpolateFile().

This means we can use standard Haxe templating, but don't have to rely on the developer
making sure the correct views exist in the view directory at runtime.